i had this whole message typed out, then realised i was writing in the brad file comment page :(((((( (Brad Patch 1.4, i think) (something about yummy spagethi code and me being at almost 500 lines, eat that corey) (oh and everything is documented so heinz can stop complaining about nothing making sense and/or being documented)
everything is pretty much done. all pop ups are done and finalised, 85% of the functions work. still having a problem with my list of patients and tests. its just a PyQt5 issue i'm having that i forgot to ask Tuxta about. besides that its just the updating details, where i'm having another PyQt5 problem. it is what it is tho. it's enough to start working on formatting it in my document and submitting draft.